# openapi.yaml
openapi: 3.1.0
info:
  title: NovaMember API
  version: 1.0.0
  description: |
    Shopify + Recharge membership management platform with AI-powered engagement scoring.
    All authenticated endpoints require a Bearer JWT obtained from `/api/auth/login` or `/api/checkout`.
  contact:
    name: NovaMember Support
    email: support@novamember.io
  license:
    name: Proprietary
servers:
  - url: https://novamember.io
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/subscriptions/upgrade:
    post:
      tags: [Subscriptions]
      summary: Upgrade or downgrade subscription plan
      description: |
        Prorates the price difference for the remaining billing period.
        Cancels the old Recharge subscription and creates a new one atomically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subscriptionId, newPlanId]
              properties:
                subscriptionId: { type: string }
                newPlanId:      { type: string, enum: [starter, growth, scale] }
      responses:
        '200':
          description: Plan changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription:   { $ref: '#/components/schemas/Subscription' }
                  proratedAmount: { type: number }
                  isUpgrade:      { type: boolean }
        '400':
          description: Already on this plan
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Subscription not found

  # ── Evals ─────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access JWT from `/api/auth/login` or `/api/checkout`
    cronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
      description: Vercel cron secret from `CRON_SECRET` env var
    rechargeHmac:
      type: apiKey
      in: header
      name: X-Recharge-Hmac-Sha256
    shopifyHmac:
      type: apiKey
      in: header
      name: X-Shopify-Hmac-Sha256
  responses:
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:   { type: string }
        code:    { type: string, enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, EXTERNAL_SERVICE_ERROR, PAYMENT_REQUIRED, INTERNAL_ERROR] }
        details: { type: object, nullable: true }
...
    Subscription:
      type: object
      properties:
        id:                     { type: string }
        plan:                   { type: string }
        price:                  { type: number }
        billingCycle:           { type: string }
        status:                 { type: string, enum: [ACTIVE, CANCELLED, PAUSED, PAST_DUE] }
        currentPeriodEnd:       { type: string, format: date-time }
        rechargeSubscriptionId: { type: string, nullable: true }
...